refactor: memberType + InterviewStatusType package 변경
