+++ backend/src/main/java/com/woowacourse/ternoko/interview/application/InterviewService.java
@@ -14,7 +14,7 @@ import com.woowacourse.ternoko.availabledatetime.domain.AvailableDateTimeReposit
 import com.woowacourse.ternoko.comment.repository.CommentRepository;
 import com.woowacourse.ternoko.common.exception.CoachNotFoundException;
 import com.woowacourse.ternoko.common.exception.CrewNotFoundException;
-import com.woowacourse.ternoko.domain.InterviewStatusType;
+import com.woowacourse.ternoko.interview.domain.InterviewStatusType;
 import com.woowacourse.ternoko.domain.member.Coach;
 import com.woowacourse.ternoko.domain.member.Crew;
 import com.woowacourse.ternoko.interview.domain.FormItemRepository;
