+++ backend/src/main/java/com/woowacourse/ternoko/login/aop/LoginMemberVerifier.java
@@ -3,7 +3,7 @@ package com.woowacourse.ternoko.login.aop;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.COACH_NOT_ALLOWED;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.CREW_NOT_ALLOWED;
 
-import com.woowacourse.ternoko.login.domain.MemberType;
+import com.woowacourse.ternoko.domain.member.MemberType;
 import com.woowacourse.ternoko.login.exception.CoachNotAllowedException;
 import com.woowacourse.ternoko.login.exception.CrewNotAllowedException;
 import lombok.RequiredArgsConstructor;
