+++ backend/src/main/java/com/woowacourse/ternoko/login/aop/MemberTypeCache.java
@@ -1,6 +1,6 @@
 package com.woowacourse.ternoko.login.aop;
 
-import com.woowacourse.ternoko.login.domain.MemberType;
+import com.woowacourse.ternoko.domain.member.MemberType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.context.annotation.RequestScope;
