+++ backend/src/test/java/com/woowacourse/ternoko/comment/application/CommentServiceTest.java
@@ -23,11 +23,11 @@ import com.woowacourse.ternoko.comment.exception.InvalidCommentMemberIdException
 import com.woowacourse.ternoko.comment.exception.InvalidStatusCreateCommentException;
 import com.woowacourse.ternoko.comment.exception.InvalidStatusFindCommentException;
 import com.woowacourse.ternoko.common.exception.MemberNotFoundException;
-import com.woowacourse.ternoko.domain.InterviewStatusType;
+import com.woowacourse.ternoko.interview.domain.InterviewStatusType;
 import com.woowacourse.ternoko.interview.application.InterviewService;
 import com.woowacourse.ternoko.interview.dto.InterviewResponse;
 import com.woowacourse.ternoko.interview.exception.InterviewNotFoundException;
-import com.woowacourse.ternoko.login.domain.MemberType;
+import com.woowacourse.ternoko.domain.member.MemberType;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
